ci: use testing storybook for interaction tests
